% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPath2.R
\name{ggplotPath2}
\alias{ggplotPath2}
\alias{ggplotPath2.list}
\alias{ggplotPath2.mts}
\alias{ggplotPath2.KFS}
\alias{ggplotPath2.default}
\title{\code{ggplot} multiple paths in one panel each with no space between panels and
a shared horizontal axis on the bottom.}
\usage{
ggplotPath2(object, ...)

\method{ggplotPath2}{list}(object, ...)

\method{ggplotPath2}{mts}(object, time, ...)

\method{ggplotPath2}{KFS}(object, ...)

\method{ggplotPath2}{default}(
  object,
  time,
  object2,
  scaley,
  logy,
  ylab,
  hlines,
  vlines = numeric(0),
  labels,
  fontsize = 10,
  color,
  linetype,
  ...
)
}
\arguments{
\item{object}{with an associated matrix to be plotted. The following
classes of objects are supported:
\itemize{
\item \code{matrix}, \code{\link{data.frame}}, \code{tibble} or \code{mts} of \code{dim = c(n, k)},
e.g., \code{k} components of a state vector at \code{n} points in time.
\code{ggplotPath} is called separately for each of the \code{k} column. Each such
call returns a \code{ggplot} object. Those \code{ggplot} objects are assembled as
separate panels into a \code{ggplot} object with a shared horizontal axis
with no space between the panels.
\item \code{KFS} (defined in package \code{KFAS}) calls
\code{ggplotPath.matrix(object$a, ...)}.
\item \code{list} with a component \code{a}, assumed to a model produced from
estimation in the \code{KFAS} package, calls
\code{ggplotPath2(KFAS::KFS(object$a), ...)}.
}}

\item{...}{optional arguments.}

\item{time}{If present, it must be a numeric vector of length equal to the
number of rows of the matrix obtained from \code{object} and fed to \code{ggplotPath}
to create the individual panels of the plot. Default depends on
\code{class(object)}:
\itemize{
\item \code{matrix}: Default \code{time = rownames(object)}.
\item \code{mts}: Default \code{time = time(object)}.
\item \code{KFS} (defined in package \code{KFAS}): Default = \code{rownames} or \code{time}
of \code{object$model$y}.
\item \code{model} (defined in package \code{KFAS}): Default = \code{rownames} or \code{time}
of \code{object$y}.
}}

\item{object2}{an optional vector or matrix-type object with the same
number of rows as \code{object} and \code{k2} columns that provide separate reference
lines to appear in the same panels as the first \code{k2} columns of \code{object}.
Obviously, \code{k2 <= k}.}

\item{scaley}{= optional numeric vector of the length \code{k} fed individually
as the scaley argument accompanying the different successive calls to
\code{ggplotPath}. If the matrix obtained from \code{object} has two columns with
names = c('level', 'slope') or c('level', 'growthRate'), then the default
for \code{scaley} = c(1000, 100). Otherwise, default = 1.}

\item{logy}{optional character vector of length \code{k} to control the scales
used in the individual panels of the plot:
\itemize{
\item \code{exp_log}: Feed \code{exp} of column \code{i} of the matrix to plot to
\code{ggplotPath(..., logy=TRUE)}.
\item \code{log}: Feed column \code{i} of the matrix to plot to
\code{ggplotPath(..., logy=TRUE)}.
\item '': Feed column \code{i} of the matrix to plot to
\code{ggplotPath(..., logy=FALSE)}.
}
The default is 'exp_log' for the first panel and '' for the rest.}

\item{ylab}{optional character vector of length \code{k} for y axis labels.}

\item{hlines}{optional list of at most \code{k} of numeric vectors (possibly with
attributes), for horizontal lines in panel \code{i}, feeding \code{hlines[[i]]} to
\code{ggplotPath(..., hlines = hlines[[i]])} for panel \code{i = 1:k}.
\verb{color='grey', lty='dotted'} unless \code{color} or \code{colour} and / or \code{lty}
are available as \code{attr(hlines[[i]], ...)}.}

\item{vlines}{optional numeric vector of locations on the \code{x} axis for
vertical lines using \code{ggplotPath(..., vlines=vlines)} with \verb{color='grey',  lty='dotted'} unless \code{color} or \code{colour} and / or \code{lty} are available as
\code{attr(vlines, ...)}.}

\item{labels}{= optional \code{\link{data.frame}} with columns
\verb{x, y, label, component}, and optionally \verb{srt, col, size}, where \code{x}, \code{y},
\code{srt}, and \code{size} are are numeric, \code{label} is character, and \code{col} are
acceptable values for \code{color} in
\code{with(labels, annotate('text', x=x, y=y, label = label, srt=srt,  color=col, size=size))}. Defaults for \code{srt}, \code{col}, and \code{size}  are 0,
'black', and 4, respectively. \code{component} is an integer in \code{1:k}.
\code{labelsi = subset(labels, component==i)} is used in
\code{ggplotPath(..., labels=labelsi)} to label panel \code{i = 1:k}.}

\item{fontsize}{optional number for legend and axes labels, used in
`ggplotPath(..., fontsize=fontsize); default = 10.}

\item{color}{optional vector or list of length '\code{k} of vectors to feed to
\code{ggplotPath(..., color=color[[i]])} for panel \code{i=1:k}. If present,
\code{length(color)} should equal \code{k}. Default is 2:1 for each panel with a
reference line provided in \code{object2} and 1 otherwise. See \code{help(ggplotPath)}
for options.}

\item{linetype}{optional vector or list of length \code{k} of vectors to feed to
\code{ggplotPath(..., linetype=linetype[[i]])} for panel \code{i=1:k}. If present,
\code{length(color)} should equal \code{k}. Default is 2:1 for each panel with a
reference line provided in \code{object2} and 1 otherwise. See \code{help(ggplotPath)}
for options.}
}
\value{
an object of class \code{\link[ggplot2:ggplot]{ggplot2::ggplot}}, which can be subsequently
edited, and whose \code{\link{print}} method produces the desired plot.
}
\description{
\code{ggplotPath2} accepts a variety of inputs. The most general is with \code{object}
= a list of sublists to be fed individually to \code{do.call(ggplotPath,..)} and
then assembled into the desired plot after eliminating the space between the
individual plots. Optional \verb{\\dots} arguments give default values for the
individual calls to \code{\link{ggplotPath}}.
}
\details{
Alternatively, the first \code{object} argument can be a matrix, \code{\link{data.frame}},
\code{tibble}, or multivariate time series (of class \code{mts}). An optional \code{Time}
argument must be a numeric vector with length equal to the number of rows of
\code{object}. If \code{Time} is provided, it overrides any values for the horizontal
axis that could be inferred from \code{\link{rownames}} or \code{\link{time}} of \code{object}. A
second optional \code{object2} can be either a vector whose length matches the
number of rows of \code{object} or a matrix or \code{ts} object similarly matching the
number of rows of \code{object}.
}
\examples{
# matrix examples 
Mat <- cbind(lvl=1:5, vel=rep(1:2, length=5), acc=sin(1:5))
Mat1 <- Mat
rownames(Mat1) <- 1951:1955
Mat2 <- as.data.frame(cbind(Mat, year=1951:1955))

# mts example 
MTS <- ts(Mat, 1951)

# Do 
Matp <- ggplotPath2(Mat)
Mat1p <- ggplotPath2(Mat1)
Mat2p <- ggplotPath2(Mat2[, 1:3], time=Mat2[, 'year']) 
MTSp <- ggplotPath2(MTS)
MTSep <- ggplotPath2(MTS, logy=c('exp_log', 'log', ''))

# list example 
List2 <- list(
    level=list('year', 'lvl'), 
    slope=list('year', 'vel'), 
    accel=list('year', 'acc'))
Mat2l <- ggplotPath2(List2, data=Mat2)

# State space / Kalman filtering model for GBR
GBR <- subset(MaddisonData, (ISO=='GBR') & !is.na(gdppc))
# model example 
growthFormula <- (log(gdppc)~ -1 + SSMbespoke(growthModel(.04, GBR$gdppc) )) 
library(KFAS)
GBR2m <-SSModel(growthFormula, GBR, H=matrix(NA) )

# NOTE: This call ignores Time 
GBRgrowthFit1 <- fitSSM(GBR2m, inits=-6, method = "BFGS", 
                        updatefn = growthUpdateFn)
# NOTE: This call currently also ignores Time; MUST BE FIXED
GBRgrowthFit1t <- fitSSM(GBR2m, inits=-6, method = "BFGS", 
                         updatefn = growthUpdateFn, Time=GBR$year)
                       
#KFS example 
GBR_KFS <- KFAS::KFS(GBRgrowthFit1$model)
GBR_KFSt <- KFAS::KFS(GBRgrowthFit1t$model)
GBR_KFSp0 <- ggplotPath2(GBR_KFS)
GBR_KFSp <- ggplotPath2(GBR_KFS$a)
GBR_KFStp <- ggplotPath2(GBR_KFSt$a)

# label the lines
ISOll1 <- data.frame(x=c(1500, 1800), y=c(2.5, 1.7), 
                 label=c('GBR', 'Napoleon'), srt=c(0, 30),
                 col=c('red', 'green'), size=c(2, 9), 
                 component=1)

GBR_KFSp1 <- ggplotPath2(GBR_KFS$a, labels=ISOll1)
                 
GBR_KFSp <- ggplotPath2(GBR_KFS, labels=ISOll1)
ISOll2 <- ISOll1
ISOll2$component <- 1:2
GBR_KFSp2 <- ggplotPath2(GBR_KFS, labels=ISOll2)

# hlines, vlines 
zero <- 0
attr(zero, color) <- 'red'
attr(zero, lty) <- 'dashed'
Hlines1 <- list(c(1,3, 10, 30), zero) 
Vlines <- c(1649, 1929, 1933, 1945)

GBR_KFSp3 <- ggplotPath2(GBR_KFS, ISOll2, hlines=Hlines1, vlines=Vlines)

}
