% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthModel.R
\name{growthModel}
\alias{growthModel}
\title{Two-dimensional growth model for \code{KFAS}}
\usage{
growthModel(
  sigma,
  y,
  a1,
  Time,
  stateNames = c("level", "growthRate"),
  Log = TRUE,
  ...
)
}
\arguments{
\item{sigma}{a numeric vector, forced to length 2 by replacing it by
\code{rep(sigma, length=2)}. The one-step transition variance for
\code{level} is \code{Q[1,1] = sigma[1]^2} and for \code{growthRate} is
\code{Q[2,2] = sigma[2]^2}. With \code{m} missing values,
\code{Q[1,1] = (m+1)*sigma[1]^2}, \code{Q[2,2] = (m+1)*sigma[2]^2}, and
\verb{Q[1,2] = Q[2,1] = sqrt((m+1)*choose(m+1,2)*sigma[1]*sigma[2]}.}

\item{y}{= optional numeric vector or \code{ts} object of length at least 2 used
to estimate the \code{a1} parameter for \code{\link[KFAS:SSModel]{KFAS::SSMcustom()}} if \code{a1} is not
supplied and for \code{Time} if it is not supplied.
\code{a1 <- (if(Log) log(c(y[1], y[2]/y[1])) else c(y[1], y[2]-y[1]))}.}

\item{a1}{= optional numeric vector of length 2 to pass to
\code{KFAS::SSMcustom}. If supplied, the numeric values of \code{y} are ignored.}

\item{Time}{= optional integer vector of times at which non-missing
observations are available. Default = \code{time(y)} if \code{y} is of class \code{ts} or
or \code{names(y)} if \code{!is.null} or \code{1:length(y)} otherwise.}

\item{stateNames}{= \code{c('level', 'growthRate')}}

\item{Log}{default = TRUE.}

\item{\dots}{optional arguments passed to \code{\link[KFAS:SSModel]{KFAS::SSMcustom()}}.}
}
\value{
a list returned by \code{\link[KFAS:SSModel]{KFAS::SSMcustom()}}.
}
\description{
\code{growthModel} returns a list returned by \code{\link[KFAS:SSModel]{KFAS::SSMcustom()}} for a model
with potentially irregularly spaced univariate observations with a
2-dimensional (\code{level}, \code{growthRate}) state.
}
\examples{
GBR <- subset(MaddisonData, (ISO=='GBR') & !is.na(gdppc))
growthMdl1 <- growthModel(.1, GBR$gdppc, Time=GBR$year)
GBRgdppc1 <- with(GBR[-1, ], ts(gdppc, year[1]))
growthMdl2 <- growthModel(c(.1, .2), GBRgdppc1)
growthMdl0 <- growthModel(.1, GBR$gdppc, a1=c(10, 1), Log=FALSE, 
                  stateNames=c('lvl', 'vel'))

}
\keyword{models}
