% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthUpdateFn.R
\name{growthUpdateFn}
\alias{growthUpdateFn}
\title{Update function to estimating a two-dimensional growth model}
\usage{
growthUpdateFn(pars, model, Time)
}
\arguments{
\item{pars}{= \code{log(variance)} of noise for \verb{[1] observations},
\verb{[2] level}, and \verb{[3] growthRate}. Forced to length 3 via
\code{c(pars, rep(tail(pars, 1), length=3-length(pars)))}. With no missing
values, the one-step transition variance for \code{level} is
\verb{Q[1,1] = exp(pars[2]} and for \code{growthRate} is \code{Q[2,2, ] = exp(pars[3])}.
With \code{m} missing values, \code{Q[1,1, ] = (m+1)*exp(pars[2])},
\code{Q[2,2] = (m+1)*exp(pars[3])}, and
\verb{Q[1,2, ] = Q[2,1] = sqrt((m+1)*choose(m+1,2)*exp(mean(pars[2:3]))}.}

\item{model}{= list assumed to have components \code{n} = third dimension of the
transition covariance array \code{Q} and
\code{state_names = c('level', 'growthRate')} as returned by \code{\link{growthModel}}.}

\item{Time}{= optional integer vector of times at which non-missing
observations are available. Default is \code{model$Time} if available or
\code{time(y)} if \code{model} has a component \code{y} or \code{names(y)} if \code{!is.null} or
\code{1:n} otherwise.}
}
\value{
a \code{model} with components \code{H} and \code{Q} updated as described.
}
\description{
\code{growthUpdateFn}
}
\details{
returns a list returned by \code{\link[KFAS:SSModel]{KFAS::SSMcustom()}} for a model
with potentially irregularly spaced univariate observations with a
2-dimensional (\code{level}, \code{growthRate}) state.
}
\examples{
GBR <- subset(MaddisonData, (ISO=='GBR') & !is.na(gdppc))
growthMdl1 <- growthModel(.1, GBR$gdppc, Time=GBR$year)
growthMdl1v0 <- growthUpdateFn(.1, growthMdl1)
growthMdl1v <- growthUpdateFn(.1, growthMdl1, Time=GBR$year)
growthMdl1v3 <- growthUpdateFn(1:3, growthMdl1, Time=GBR$year)

growthFml <- (log(gdppc)~ -1 + SSMbespoke(growthModel(.1, GBR$gdppc) )) 
library(KFAS)
growthSSmdl <-SSModel(growthFml, GBR, H=matrix(NA) )
growthSSmdl1 <- growthUpdateFn(.1, growthSSmdl)

}
\keyword{models}
