% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaddison.R
\name{plotMaddison}
\alias{plotMaddison}
\title{Plot selected countries}
\usage{
plotMaddison(
  ISO,
  y,
  lty,
  col,
  lwd,
  data = MaddisonData::MaddisonData,
  scaley = 1000,
  logy = TRUE,
  Size = c(text = 14, axes = 11, ISO = 14),
  ISOlabelLoc
)
}
\arguments{
\item{ISO}{either NULL to select all the data in \code{MaddisonData} or a
character vector of \code{ISO} codes used in the Maddison project.}

\item{y}{name of column in \code{data} to plot. Default = 1.}

\item{lty, col, lwd}{line types, colors, and widths for the different lines,
one for each ISO. Defaults for \code{lty} and \code{col} = 1:length(\code{ISO}). Default
for \code{lty}  = 1.}

\item{data}{\code{\link{data.frame}} or \code{\link[tibble:tibble]{tibble::tibble}} with first two columns
being \code{ISO} and \code{year} and \code{y} being the name of another column.}

\item{scaley}{factor to divide y by for plotting. Default = 1000, e.g., so
numbers in thousands are plotted in millions.}

\item{logy}{logical: if \code{TRUE}, y axis is on a log scale; default = \code{TRUE}.`}

\item{Size}{size of axis text labels, tick labels and ISO code in a named
3-vector with default = c(axLab=14, tickLabs=11, ISO=14).}

\item{ISOlabelLoc}{~legend:
If a character vector of length 1 or a numeric vector of length 2, it is
used as the \code{legend.position} argument in \code{theme}.
Otherwise, it should be a matrix with rownames = ISO codes and columns \code{x}
and \code{yAdj}: The algorithm finds the closest points
above and below \code{x} and averages both \code{x} and \code{y} for the 1, 2, or 3 points
found bracketing \code{x}, then multiplies the average \code{y} by \code{yAdj} and plots
the \code{ISO} label there with \code{col} specified for that \code{ISO} code. If
\code{ISOlabelLoc} is missing or \code{ISO} is not found among the rownames of
\code{ISOlabelLoc}, that line is not labeled.
Default = c(.1, .9) for top left inside the plot.}
}
\value{
an object of class \code{\link[ggplot2:ggplot]{ggplot2::ggplot}}, which can be subequently
edited, and whose \code{\link{print}} method produces the desired plot.
}
\description{
\code{plotMaddison} plots y vs. year with a separate line for each ISO country
code.
}
\examples{
(GBR_USA <- plotMaddison(c('GBR', 'USA'))) # GBR, USA
GBR_USA+ggplot2::coord_cartesian(xlim=c(1500, 1850)) # for only 1500-1850 
GBR_USA+ggplot2::coord_cartesian(xlim=c(1600, 1700), ylim=c(7, 17)) 

ISOll <- matrix(c(1500, 1750, 1.4, .7), 2, dimnames=
            list(c('GBR', 'USA'), c('x', 'yAdj')))
(GBR_USA_ <- plotMaddison(c('GBR', 'USA'), ISOlabelLoc=ISOll) ) 
# label the lines

}
\keyword{plot}
